Support for wxWidgets 2.9.4. Requires changes in the environment variables (WXWINVER and WXWINDIR instead of WXWIN28), so run trunk/Build/Installer/update.bat to update your variables.

+++ Tools/EditorBase/WXScintilla/scintilla/src/ScintillaBase.cxx
@@ -96,7 +96,7 @@ bool ScintillaBase::Command(int cmdId) {
 	if (cmdId >= idcmdUserDefined) {
 
 		cmdId -= (int) idcmdUserDefined;
-		assert(cmdId < contextUserActions.size());
+		assert(cmdId < (int) contextUserActions.size());
 
 		contextUserActions[cmdId].second.Action(contextMenuX, contextMenuY);	// Invoke respective user action.
 		return true;
